refactor: move shell and env probing helpers into orchestrator
