+++ crates/cella-cli/src/commands/env_cache.rs
@@ -1,217 +1,3 @@
-use std::collections::HashMap;
-
-use tracing::debug;
-
-use cella_backend::{ContainerBackend, ExecOptions, FileToUpload};
-use cella_env::platform::DockerRuntime;
-
-/// Compute the per-user cache path for the probed environment.
-///
-/// Stores under `$HOME/.cella/probed-env.json` so it survives container
-/// restarts (unlike `/tmp` which can be cleared).
-fn cache_path(user: &str) -> String {
-    let home = cella_env::claude_code::container_home(user);
-    format!("{home}/.cella/probed-env.json")
-}
-
-/// Read the cached probed environment from a running container.
-///
-/// Returns `None` if the cache file doesn't exist or can't be parsed
-/// (graceful fallback -- never errors).
-pub async fn read_probed_env_cache(
-    client: &dyn ContainerBackend,
-    container_id: &str,
-    user: &str,
-) -> Option<HashMap<String, String>> {
-    let path = cache_path(user);
-    let result = client
-        .exec_command(
-            container_id,
-            &ExecOptions {
-                cmd: vec!["cat".to_string(), path],
-                user: Some(user.to_string()),
-                env: None,
-                working_dir: None,
-            },
-        )
-        .await
-        .ok()?;
-
-    if result.exit_code != 0 {
-        debug!("No probed env cache found (exit {})", result.exit_code);
-        return None;
-    }
-
-    let env: HashMap<String, String> = serde_json::from_str(result.stdout.trim()).ok()?;
-    debug!("Read {} probed env vars from cache", env.len());
-    Some(env)
-}
-
-/// Probe the user's environment and cache it inside the container.
-///
-/// Returns the probed environment, or `None` if probing is disabled or fails.
-pub async fn probe_and_cache_user_env(
-    client: &dyn ContainerBackend,
-    container_id: &str,
-    user: &str,
-    probe_type: &str,
-    shell: &str,
-) -> Option<HashMap<String, String>> {
-    let env = run_env_probe(client, container_id, user, probe_type, shell).await?;
-    write_env_cache(client, container_id, user, &env).await;
-    Some(env)
-}
-
-/// Execute the environment probe command and parse the output.
-async fn run_env_probe(
-    client: &dyn ContainerBackend,
-    container_id: &str,
-    user: &str,
-    probe_type: &str,
-    shell: &str,
-) -> Option<HashMap<String, String>> {
-    let probe_cmd = cella_env::user_env_probe::probe_command(probe_type, shell)?;
-
-    debug!("Running userEnvProbe ({probe_type}) with {shell}: {probe_cmd:?}");
-
-    let result = client
-        .exec_command(
-            container_id,
-            &ExecOptions {
-                cmd: probe_cmd,
-                user: Some(user.to_string()),
-                env: None,
-                working_dir: None,
-            },
-        )
-        .await
-        .ok()?;
-
-    if result.exit_code != 0 {
-        debug!(
-            "userEnvProbe failed (exit {}): {}",
-            result.exit_code,
-            result.stderr.trim()
-        );
-        return None;
-    }
-
-    let env = cella_env::user_env_probe::parse_probed_env(&result.stdout);
-    if env.is_empty() { None } else { Some(env) }
-}
-
-/// Write the probed environment to a cache file inside the container.
-///
-/// Creates the `~/.cella/` directory if it doesn't exist.
-async fn write_env_cache(
-    client: &dyn ContainerBackend,
-    container_id: &str,
-    user: &str,
-    env: &HashMap<String, String>,
-) {
-    let Some(json) = serde_json::to_string(env).ok() else {
-        return;
-    };
-
-    let path = cache_path(user);
-
-    // Ensure the parent directory exists
-    let home = cella_env::claude_code::container_home(user);
-    let dir_path = format!("{home}/.cella");
-    let _ = client
-        .exec_command(
-            container_id,
-            &ExecOptions {
-                cmd: vec!["mkdir".to_string(), "-p".to_string(), dir_path],
-                user: Some(user.to_string()),
-                env: None,
-                working_dir: None,
-            },
-        )
-        .await;
-
-    let cache_file = FileToUpload {
-        path,
-        content: json.into_bytes(),
-        mode: 0o644,
-    };
-
-    if let Err(e) = client.upload_files(container_id, &[cache_file]).await {
-        debug!("Failed to cache probed env: {e}");
-    } else {
-        debug!("Cached {} probed env vars", env.len());
-    }
-}
-
-/// If `SSH_AUTH_SOCK` is not already present in `env`, detect the runtime and
-/// check whether the well-known socket path exists inside the container.
-/// Appends `SSH_AUTH_SOCK=<path>` if found.
-pub async fn ensure_ssh_auth_sock(
-    client: &dyn ContainerBackend,
-    container_id: &str,
-    user: &str,
-    env: &mut Vec<String>,
-) {
-    if env.iter().any(|e| e.starts_with("SSH_AUTH_SOCK=")) {
-        return;
-    }
-
-    let runtime = cella_env::platform::detect_runtime();
-    let socket_path = if matches!(
-        runtime,
-        DockerRuntime::DockerDesktop | DockerRuntime::OrbStack
-    ) {
-        "/run/host-services/ssh-auth.sock"
-    } else {
-        "/tmp/cella-ssh-agent.sock"
-    };
-
-    let result = client
-        .exec_command(
-            container_id,
-            &ExecOptions {
-                cmd: vec![
-                    "test".to_string(),
-                    "-e".to_string(),
-                    socket_path.to_string(),
-                ],
-                user: Some(user.to_string()),
-                env: None,
-                working_dir: None,
-            },
-        )
-        .await;
-
-    if let Ok(r) = result
-        && r.exit_code == 0
-    {
-        debug!("SSH_AUTH_SOCK fallback: found {socket_path}");
-        env.push(format!("SSH_AUTH_SOCK={socket_path}"));
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn cache_path_root_user() {
-        let path = cache_path("root");
-        assert!(path.contains("root"));
-        assert!(path.ends_with("/.cella/probed-env.json"));
-    }
-
-    #[test]
-    fn cache_path_regular_user() {
-        let path = cache_path("vscode");
-        assert!(path.contains("vscode"));
-        assert!(path.ends_with("/.cella/probed-env.json"));
-    }
-
-    #[test]
-    fn cache_path_custom_user() {
-        let path = cache_path("devuser");
-        assert!(path.contains("devuser"));
-        assert!(path.ends_with("probed-env.json"));
-    }
-}
+pub use cella_orchestrator::env_cache::{
+    ensure_ssh_auth_sock, probe_and_cache_user_env, read_probed_env_cache,
+};
